[CRES-61] refactor: 폴더이름 `query` -> `queries`로 변경 #59
